Revert "Move metrics to their own files"
This reverts commit ed19f51a7af6d4965fa0108755ee94364704bbd3.
