--- city_metrix/metrics/__init__.py
@@ -1,5 +1,0 @@
-from built_land_without_tree_cover import built_land_without_tree_cover
-from built_land_with_low_surface_reflectivity import built_land_with_low_surface_reflectivity
-from built_land_with_high_land_surface_temperature import built_land_with_high_land_surface_temperature
-from mean_tree_cover import mean_tree_cover
-from urban_open_space import urban_open_space
